evaluation/requirements.txt
ragas
langchain-google-genai
pandas
python-dotenv
openai
